refactor repo for details screen
